powershell terminal has already had ANSI sequence code supported

+++ tamer/cc/ffi/nested/seed.c
@@ -2,7 +2,7 @@
 #include <time.h>
 
 __lambda__ int random_seed() {
-    srand(time(NULL) + clock());
+    srand((unsigned int)(time(NULL)) + clock());
 
     return rand() % 64;
 }
